move _hrnow to utils
