+++ lib/tracing/trace.js
@@ -3,7 +3,8 @@ const cds = require('@sap/cds')
 const otel = require('@opentelemetry/api')
 const { SpanKind, SpanStatusCode, ROOT_CONTEXT } = otel
 const { SemanticAttributes } = require('@opentelemetry/semantic-conventions')
-const { hrTimeToMilliseconds } = require('@opentelemetry/core')
+
+const { _hrnow } = require('../utils')
 
 const MASK_HEADERS = (cds.env.log.mask_headers || ['/authorization/i', '/cookie/i']).map(s => {
   const parts = s.match(/\/(.+)\/(\w*)/)
@@ -12,26 +13,9 @@ const MASK_HEADERS = (cds.env.log.mask_headers || ['/authorization/i', '/cookie/
 })
 
 const HRTIME = cds.env.requires.telemetry.tracing._hrtime && cds.env.requires.telemetry.tracing._hrtime !== 'false'
-const now = Date.now()
-const hrTimeInMS = Number(`${hrTimeToMilliseconds(process.hrtime())}`.split('.')[0])
-const diff = now - hrTimeInMS
-const EPOCH_OFFSET_S = Number(`${diff}`.slice(0, -3))
-const EPOCH_OFFSET_MS = Number(`${diff}`.slice(-3) + '000000')
 
 // attach a hr time to incoming requests for later adjustment of span start time in _getParentSpan()
 if (HRTIME) cds.on('listening', ({ server }) => server.on('request', req => (req.__hrtime = _hrnow())))
-
-// returns [seconds, nanoseconds] since unix epoch
-function _hrnow() {
-  const hrtime = process.hrtime()
-  let s = hrtime[0] + EPOCH_OFFSET_S
-  let ns = hrtime[1] + EPOCH_OFFSET_MS
-  if (ns >= 1000000000) {
-    s++
-    ns -= 1000000000
-  }
-  return [s, ns]
-}
 
 function _getParentSpan() {
   if (!cds.context) return
