+++ lib/utils.js
@@ -4,6 +4,7 @@ const LOG = cds.log('telemetry')
 const fs = require('fs')
 
 const { DiagLogLevel } = require('@opentelemetry/api')
+const { hrTimeToMilliseconds } = require('@opentelemetry/core')
 const { Resource } = require('@opentelemetry/resources')
 const { SemanticResourceAttributes } = require('@opentelemetry/semantic-conventions')
 
@@ -129,6 +130,24 @@ function hasDependency(name) {
   return !!PKG.dependencies[name]
 }
 
+const now = Date.now()
+const hrTimeInMS = Number(`${hrTimeToMilliseconds(process.hrtime())}`.split('.')[0])
+const diff = now - hrTimeInMS
+const EPOCH_OFFSET_S = Number(`${diff}`.slice(0, -3))
+const EPOCH_OFFSET_MS = Number(`${diff}`.slice(-3) + '000000')
+
+// returns [seconds, nanoseconds] since unix epoch
+function _hrnow() {
+  const hrtime = process.hrtime()
+  let s = hrtime[0] + EPOCH_OFFSET_S
+  let ns = hrtime[1] + EPOCH_OFFSET_MS
+  if (ns >= 1000000000) {
+    s++
+    ns -= 1000000000
+  }
+  return [s, ns]
+}
+
 function _require(name) {
   name = Array.isArray(name) ? name[0] : name
   try {
@@ -146,5 +165,6 @@ module.exports = {
   getDynatraceCredentials,
   getCloudLoggingCredentials,
   hasDependency,
+  _hrnow,
   _require
 }
